Fix retry not working in LDAP RetryPool.
The `RetryPool.search()` function is supposed to try twice in the case
of LDAP errors. This allows us to try another server if the first one
fails, or to try the same server twice in the case there is a network
error or something.

The exception handling code in the search function failed to catch the
exception because we end up with two versions of the
`ldap3.core.exceptions.LDAPExceptionError` class. There is one version
that is used by the `ldap3` library when throwing the exception, and a
different version that we catch. The latter version is a patched
version created by eventlet.

This patch works around this problem by catching all exceptions.